+++ src/MovableObject.cpp
@@ -214,7 +214,7 @@ namespace core
 
 	void MovableObject::ComputeDiameter()
 	{
-		float maxDiameter = 0;
+		double maxDiameter = 0;
 		for (auto child : _children)
 		{
 			maxDiameter = max(maxDiameter, child->GetDiameter());
